Extract string_dependencies into its own files
Because we can.
