+++ jbmc/unit/solvers/strings/string_refinement/dependency_graph.cpp
@@ -10,6 +10,7 @@ Author: Diffblue Ltd.
 #include <testing-utils/use_catch.h>
 
 #include <java_bytecode/java_types.h>
+#include <solvers/strings/string_dependencies.h>
 #include <solvers/strings/string_refinement_util.h>
 #include <util/arith_tools.h>
 #include <util/std_expr.h>
